pad sequences to a multiple of 8 for attn calculations (xformers)

+++ pretrain_stage.py
@@ -98,7 +98,7 @@ def main(config_dir: str):
     _ = model_config.pop("position_embedders")
     
     model = OmLLM(position_embedders=position_embedders, **model_config)
-    model = set_om_dtypes(model, torch.bfloat16)
+    model = set_om_dtypes(model, torch.float16)
     
     # Determine training stage
     for stage_ix in range(1, num_stages+1):
